+++ core/src/lib.rs
@@ -57,7 +57,7 @@ pub use program::{Program, ProgramInfo};
 
 mod kernel;
 pub use kernel::Kernel;
-pub use miden_crypto::{EMPTY_WORD, ONE, WORD_SIZE, Word, ZERO};
+pub use miden_crypto::{EMPTY_WORD, ONE, WORD_SIZE, Word, ZERO, word::LexicographicWord};
 pub mod crypto {
     pub mod merkle {
         pub use miden_crypto::merkle::{