+++ processor/src/host/advice/inputs.rs
@@ -1,4 +1,4 @@
-use alloc::vec::Vec;
+use alloc::{sync::Arc, vec::Vec};
 
 use miden_core::{
     AdviceMap, Felt, Word,
@@ -24,9 +24,9 @@ use miden_core::{
 #[cfg(not(feature = "testing"))]
 #[derive(Clone, Debug, Default, PartialEq, Eq)]
 pub struct AdviceInputs {
-    stack: Vec<Felt>,
-    map: AdviceMap,
-    store: MerkleStore,
+    pub(super) stack: Vec<Felt>,
+    pub(super) map: AdviceMap,
+    pub(super) store: MerkleStore,
 }
 
 impl AdviceInputs {
@@ -115,23 +115,13 @@ impl AdviceInputs {
     }
 
     /// Fetch a values set mapped by the given key.
-    pub fn mapped_values(&self, key: &Word) -> Option<&[Felt]> {
+    pub fn mapped_values(&self, key: &Word) -> Option<&Arc<[Felt]>> {
         self.map.get(key)
     }
 
     /// Returns the underlying [MerkleStore].
     pub const fn merkle_store(&self) -> &MerkleStore {
         &self.store
-    }
-
-    // DESTRUCTORS
-    // --------------------------------------------------------------------------------------------
-
-    /// Decomposes these `[Self]` into their raw components.
-    #[allow(clippy::type_complexity)]
-    pub(crate) fn into_parts(self) -> (Vec<Felt>, AdviceMap, MerkleStore) {
-        let Self { stack, map, store } = self;
-        (stack, map, store)
     }
 }